fix: fix bouncer seeder to fit convention
