+++ src/database/seeders/DatabaseSeeder.php
@@ -14,8 +14,8 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call([
-            UserSeeder::class,
             BouncerSeeder::class,
+            AdminSeeder::class,
         ]);
     }
 }
